Fixes and Enhancements (see detailed commit message)
Fixes:
- Handle setting of component and module prototypes differently. Fixes
error about the prototype not being writable (es6 related?)

Enhancements:
- Clear the view-ref content; allows for the template to by dynamically
changed by the parent, otherwise it retains the previous template
content
- Declare the CommonModule and FormsModule on the imports declaration,
rather than simply concatenating imports all the time.
- This makes it possible for the parent component to declare its own
module that exports child components without fear of duplicate
components in the dynamically re-created module mdClass. This is a
use-case that is important if there are dynamic components that require
other components that the component-outlet cannot anticipate
(anticipating the CommonModule and FormsModule is reasonable, but not
enough by itself).
- Guidance on using the imports and declaring a shared module of imports
will be included in the future documentation of component-outlet.