Fix component / factory memory leak (#8)
* Fix component / factory memory leak

After the dynamically created component is deleted from the DOM, the component and its factories are now destroyed / cleared.

BREAKING CHANGES:

The component class no longer inherits from the context. This (somehow) leads to a memory leak, because the context is always referenced by the class' constructor. Thus, its no longer possible to directly reference the context in the template:

{{contextMember}}  //not possible anymore
{{context.contextMember}}

* Check for null in ngDestroy